let us point extension root to some other place
When using --git-root to a local path on your laptop, the script might
not be able to find extensions when they have been fetched out in a
different path.  The optional --git-root-ext let you specify where the
extensions should be fetched from.

Example:

 ./make-release.py --git-root /srv/ --git-root-ext ~/projects/mw

With core in /srv/core and extensions in ~/project/mw/extensions.

Change-Id: I4101daae85e1283aa07bf18761b37a1a0b5014ab