Change Collection to a deploy branch
Master is not currently working but it doesn't make a huge amount
of sense to revert the offending commit there because it's working
for some other code... so! reverted it in this branch and as soon
as I figure out why it breaks production mwlib I'll change back
to normal deployments.

Bug: 58151
Change-Id: I944b8406465ea4e5697f3969b8fc91135b6bc8f4