make-release: Fetch composer dependencies for core 1.25+
Starting with MediaWiki core 1.25, tarballs will need to
include dependencies specified in core's composer.json file.

This runs `composer install --no-dev` inside the checkout
of core, and it is included in the tarball.

A --composer argument can be passed to specify the location
of a `composer` executable, otherwise it assumes that
composer is in your path.

Tested that a snapshot build of master did include the vendor/
directory, and creating a 1.19.23 tarball did not.

Bug: T74726
Change-Id: I59b135d659eeed71af6929a62c5aff8e18f85931