make-wmf-branch: Move list of branched extensions/skins to a JSON file
The list of branched extensions, skins, and submodules are moved to a
JSON file so other tools can read them without having to execute PHP
code.

Set up phpunit for this repo through the "composer test" entry point to
lint the new JSON file.

At the same time I did a little cleanup of __DIR__ in the code.

Bug: T100817
Change-Id: Ia5c5f9da93d9cc296a73ca402f49f459be79d517