Use Semantic versioning format for deploy notes
Since the deployment cycle for release 1.27, wmf versions uses a semver
format. Instead of:
wmf/1.26wmf1
it now uses
wmf/1.27.0-wmf.1

Change the required things in make-deploy-notes to work with both, the new
and the olf format (the handling of the old format is kept for the time of the
change to the new format).

Bug: T67306
Change-Id: I79a225d69b818a1b483e33f1b46a77400b39663a