Add flag to resume branching from named extension
Adds the [-c|--continue-from] optional flag to `make-wmf-branch`. Setting
this flag to the name of an extension in the `config.json` file will
resume cutting new branches on branched extensions from that extension
forward.

Currently, if branching of extensions fails on a particular extension
the workflow is:

0. Figure out why branching that extension failed, fix it.
1. Delete the extension names from `make-wmf-branch/config.json`
   that occur before the extension that failed.
2. Re-run `make-wmf-branch/make-wmf-branch`.
3. Checkout the new branch of mediawiki-core.
4. Add submodules for all the extensions you deleted from the config in
   step 1.
5. Add a new patch for the new branch of core in gerrit, +2, and merge.

New workflow would be:

0. Figure out why branching that extension failed, fix it.
1. Run `make-wmf-branch/make-wmf-branch -n <new> -o <old> -c
   <failed-extension>`.

Change-Id: Icf1f2f9a98dcd390d8a187ce11ebe6b3e2219204