Update specialExtensions submodule add
Since $this->specialExtensions is an assoc. array, the check:

    in_array( $name, $this->specialExtensions )

doesn't work as the code assumes. This patch fixes that logic. This
patch also adds specialExtension submodules using the branch specified
in config.json rather than checking out the tip of that branch. The
benefit is that gerrit will automatically bump the submodule if the tip
of a specialExtension branch updates.

Change-Id: Ic3f630e9e4869f35917185c25bf5fc302c9af29b