Fix missing commits in deploy notes
The full_git_log_as_array function loops throught the output of git log
and pushes onto a $history array whenever there is a new line that
begins with "commit".

This is problematic since the oldest commit, i.e. the last commit
shown in git log output, is never pushed into $history. This patch
pushes the contents of $commit, which should contain the oldest commit
in log output if it is not empty, onto $history before returning the
$history array.

Change-Id: I7bc1db6bb6d775adc4cc2d82a6958fcc2a110087