make-wmf-branch: Stop amending commits for branches with submodules
In Ia31a9f1d, we tried to fix submodule support in branching so they
get branched properly as well. To do this, we had to make a commit
with some submodules being committed. To avoid noise and roll it into
the prior commit we'd done for .gitreview, we'd push the resulting
single commit to Gerrit.

Then I come along in I3ba2bb8f to stop mangling .gitreview on all
the extension branches because it was superfluous. However, that
meant that the *second* commit from before was now amending the
*last* commit prior to branching.

Confused yet?

Bug: T175324
Change-Id: I8a9c1a7094cb5f7be36c101498ddd7107522c2d6