make-wmf-branch: Stop committing weird sub-submodule branching
It's ok to keep creating these branches, but the logic for adding
them as auto-tracked submodules is fragile and has broken several
times over the years.

The drawback? SWAT deploys for the sub-submodules require manual
backports. This sucks, but it sucks less than the tech debt and
isn't any worse than how we used to do things before auto-updating
submodules were used.

Change-Id: I00301d8ea2d003b8d0668a8db998d3d1df34e195