Drop Modern/CologneBlue from tarball bundle
Modern is not modern. CologneBlue is outdated.

According to WikiApiary, 253 wikis have Modern as their default skin,
and 11 have CologneBlue set as default. For comparison, MonoBook, the
former MediaWiki default skin, has 1998 wikis still using it as their
default.

Wiki sysadmins will still be able to install the skin manually if they
want, it will just no longer be provided in the tarball.

Change-Id: I72ec6ff3e5727b2859ee558b0b70d94b8480e9f2