Move to `git archive` like model for MediaWiki releases
Uses <https://github.com/Kentzo/git-archive-all> to work around the
limitation of git-archive not supporting submodules.

Usage:
 ./makerelease2.py ~/path/to/mw-core 1.31.1 --previous 1.31.0

This will only archive whatever is checked into git (including
submodules). Exclusions are also managed in git via the .gitattributes
file.

To handle a security release, the patches should be applied to the local
git tree and tagged.

This fixes a lot of different wishlist items over the years:
* T73379: Download old tarballs to diff against instead of lazily
  recreating them.
* T180522: You only need one checkout of MediaWiki core.
* T180532: Mostly fixed with the fix for T73379 and the fact that vendor
  is a submodule.
* T949: Implemented in check_wg_version()

Bug: T199467
Bug: T73379
Bug: T180522
Bug: T181116
Bug: T949
Change-Id: I320b1a149af2999030fb3b4c4ad7e0cfa21c6f63