make-wmf-branch: branch special_extensions
Currently we have 1 special_extension in config.json (CentralNotice).
Rather than create new branches of this extension every month, we add
the wmf_deploy branch of this repo as a submodule. The result of this is
that whenever CentralNotice SWAT deploys in the middle of the week a
backport to the single wmf_deploy branch bumps the CentralNotice
submodule for both wmf branches in production. This causes confusion.

Sometimes a SWAT deployer will fetch and deploy CentralNotice submodule
for one wmf branch, but miss the other wmf branch. The next time someone
pulls a backport for SWAT in the other deployed wmf branch, they pull
down unexpected changes for CentralNotice.

In general, we should avoid special cases; however, in this instance the
special case can be handled at branching rather than at deployment time.
That is, we can create a new wmf/VERSION branch each week for
CentralNotice, except rather than create that branch from master, we
create that branch from wmf_deploy. This allows all workflows for
CentralNotice to stay the same and it simplifies the SWAT workflow.

Bug: T136904
Change-Id: I376b5d0fc406a8f7557c2c9a946ef2859693a86e