make-deployment-calendar: Create wikitech:Deployments
Add a python script that creates the deployment calendar page on
Wikitech from a config in JSON. This has a few advantages:

1. No longer working from a copy-and-paste template
2. Canonical place to update deployment schedule
3. Programatic generatation vs copy-paste + find-replace

My hope is that we can put this information in a Jenkins job and
eliminate the toil of creating this page all together. The missing
pieces of that mostly relate to not having a canonical schedule for
deployments; maybe we can rig something up with phabricator calendar?

Bug: T114488
Change-Id: I1fb22df24b368414c186ad8818ecf11eda49527d