make-release: Make the branch commit explain why, not what
The main rule of our git commit message coding standards is to
justify the change you're making in a way that is identifiable
just by reading the commit title, and not to laboriously spell
out what you're doing. Anyone can look at the commit's content
and see that it's adding extensions (and skins and vendor, not
mentioned) and updating Defines.php, but it's irritating, when
looking in the git log on the deployment server, to parse this
former message, so I'm changing it to one that's a bit better.

Change-Id: I61c60ecf6de01e70ad83a6eb9338e5fbba07ee89