fix(deploy-promote): add set -eu, fix things found by shellcheck
set -u means undefined variables are an error. Define two variables in
main. set -e means if a command fails (returns a non-zero exit code),
bash terminates the script. This will force us to check for failures,
but there are no changes for that yet, because we have no test
environment in which to try this. Train conductors will hate me.

If this blocks the train, remove the set -e and re-run. Then report
problem and we'll fix.

Also ran shellcheck and fixed things it found. Mostly quotes.

Change-Id: I1da215446c20feddcbe7be36622cc96748f011c4