deploy-promote: assert repo is up-to-date
Occasionally, we'll make important changes to this script. If we're not
running the latest version of this script then we could end up pushing
to the wrong remote or the right remote with the wrong parameters. This
is intended to assert that the local repo from which you are running
deploy-promote is up-to-date with the canonical upstream (which is
hardcoded here).

Change-Id: I674e9cbeb3d6efba7db67e13a26d94dcc533c989