task-series: custom.release.date is an integer
Kept getting the error message:

    13:29:18 blockers failed: <APIError> ERR-CONDUIT-CORE: Exception
       when processing transaction of type "custom.release.date": Error
       while reading "value"
           : Expected integer, got something else.

When trying to generate new train tasks. This seems to fix it, not sure
what changed in the API. I have a vague memory of Mukunda and I fiddling
with this in order to make "custom.release.date" searchable, but I'm not
sure if this is an effect of that or not.

Change-Id: I34420b10a2b4e4027facad38a1fc62d023ef9c4c