deploy-promote: Don't specify path to scap binary
There are many calls to scap in the deploy-promote script. Most of
them are unqualified (meaning that PATH is used to find the
executable).  However, there were two outliers that specified
/usr/bin/scap.  This comment changes them to rely on PATH as well.
This is to make it feasible to use a development install of scap that
may be in a different location.

Change-Id: Ie3f237b3bcd8175697a434c0e7faac53507278d0