deployment calendar: Fully automated deployment calendar
update-wiki uses mwparserfromhell and pywikibot to remove last week from
the calendar and add next week. It also adds the previous week to the
archive. The idea is that rather than manually running this on Monday of
each week I can schedule it to run on toolforge or in jenkins and stop
worrying about the deployment calendar all together!

This has been working for a few weeks...now just need to make it pass
tests.

Change-Id: I9582789ea0b29f099651c510a6909bf6b7f36e87