Calendar: Fix automation for new format
The last time the DeploymentCalendarTool updated the calendar it put the
new calendar right under the heading for the last week. This is because
we insert *after* the last "week of" heading on the page, but *before*
any other headings; however, the new deployment calendar format uses
headings for days as well.

This is a quick fix to get the automation working again. There may be a
better way to do this in mwparserfromhell, e.g., only insert after
"week of" headings and any subheadings but before the next top-level
heading. This fixes the immediate problem though.

Change-Id: I91c064faedd26571bf6eb6b3c8f48d11d874934c