build-mv-image: Handle mid-week L10N rebuilds gracefully
Patches are always freshly reapplied during auto-stage.  When a patch
affects l10n source files, all CDB files are regenerated (2.1GB).
This happens for every backport until the patch goes away.  If a patch
survives multiple train branches, the problem is compounded. [1]

As backports are made, we end up with the most recent image being
composed of an ever increasing number of 2 or 4GB layers.

Assuming all involved nodes are usually running the latest image, this
layering is optimal in terms of data transfer.  However, the
intermediate layers are still required to be stored on the target
nodes even though their contents are shadowed by the most recent
layer. By the end of the week someone running "docker image ls" might
find a 77GB mediawiki-multiversion image on all nodes.

To keep this from happening this commit revives the "estimated
transfer percentage" logic in image_is_suitable_as_base.  This checks
to see if a large proportion of data would need to be rsync'd to the
candidate base image.  If so, don't use the image as a base.  The
threshold is hard-coded to 25, meaning if more than 25% of the staged
/srv/mediawiki needs to be rsync'd, it will fall back to
multiversion-base.

In the case of l10n patches, this means that while they're alive,
backports will result in large (~4-8GB) images, but their size will
remain stable and storage for older images will be reclaimable.

Footnotes

[1] This is a problem that needs to be studied for solutions.  This
commit is focused on a reasonable workaround to use in the meantime.

Change-Id: If23681fc1bcfdb24a87a5c59a22c429a0b6c1d06