Use the remote origin url for Gerrit/git heuristic
Relying on `/etc/wikimedia-cluster` to determine whether to push commits
in a Gerrit Review `refs/for/` fashion vs. a direct git push couples
this logic to the environment. The git remote origin URL is a better
heuristic as it's closer to the inference being made (is this a Gerrit
or simple git server remote).

Change-Id: If105605934283c4fa42dbbe9eccdf3b33461bdc5