auto-stage: Honor MEDIAWIKI_CONFIG_COMMIT and MEDIAWIKI_COMMITS
auto-stage now checks for environment variables
MEDIAWIKI_CONFIG_COMMIT and MEDIAWIKI_COMMITS and uses them to
determine which commits of operations/mediawiki-config and
mediawiki/core to stage.  Each variable defaults to the tip of the
respective branch.

MEDIAWIKI_COMMITS, if supplied, must be a JSON object whose keys are
branch names (e.g., wmf/1.38.0-wmf.7) and values are commit
references.  If no matching entry is found for a given branch, the tip
of the branch is used.

Change-Id: I22458c78f957ad03c1a018fb0eed7bb7cecd1fa8