make-release: hardened branch existence detection
delete_branch handled the non existence of a branch when deleting.
However when converting to a tag we retrieve the branch information
which would raise if it does not exist.  The code added to handle the
tag forgot to adjust the part deleting the branch.

Always retrieve the branch information, whether we want to convert to
tag or not. When the branch does not exist, return False immediately and
raise on error.

Add tests to cover the error handling when a branch does not exist or
there is an error retrieving the informations.

A side effect is running `branch.py` with `--noop` would check all
branches existences but it is not doing any write so I think it is in
the spirit of doing no operations.

Bug: T303828
Change-Id: I704f063535fb8a6a45f976f51055d7147b241193