phab-train-blocker-tasks: train blocker script
This is the script I have been using to generate train blockers.

It's based on the work Mukunda did making this a scap plugin that was
removed in f52d85a16bd8eb324cd8103ad519627fa782e6cc. The only work I did
was to remove the dependency on scap.

Change-Id: I3825a3f46da9f1232eff4708d548887cb33c7c88