Calendar: empty schedule should be ok
An empty schedule currently throws a runtime error. This is creating a
problem for next week, as we've removed all the special scheduling
considerations for weeks where there is no train. We may have other
schedules like this.

The intention of this check was to error out if we pass a schedule that
does not exist, NOT if we pass a schedule that exists, but adds nothing
to the calendar.

This updates the check to only throw a runtime error in the case where
the schedule name does not exist as a key in the yaml/json.

Change-Id: I51add4fd79a18795f6d9c22018b9f5135791d658