Remove new-train-version logic
The new-train-version script attempted to detect start-of-week train
situations where we want to perform a fresh, non-incremental
multiversion image build.  However, the detection logic also triggered
when rolling back a train, which is very undesirable.  This logic has
been removed and going forward callers of build-mv-image must set
FORCE_FULL_BUILD=true in the environment to insist on a full build.

Bug: T325576
Change-Id: I9f13c35f43ce879cc4fb7caec59f49370d078e57